requirements.txt
flask==3.1.3
mysql-connector-python==9.6.0
waitress==3.0.2
